+++ eBookStore/Pages/Publishers/Delete.cshtml.cs
@@ -1,5 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using System.Net;
+using System.Net.Http.Headers;
 using System.Text.Json;
 
 namespace eBookStore.Pages.Publishers
@@ -18,6 +20,14 @@ namespace eBookStore.Pages.Publishers
 
         public async Task<IActionResult> OnGetAsync(int id)
         {
+            var token = Request.Cookies["Token"];
+
+            if (string.IsNullOrEmpty(token))
+            {
+                return Unauthorized(); // Or redirect to login page
+            }
+            // Add the Authorization header with Bearer token
+            _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
             var response = await _httpClient.GetAsync($"Publishers/get-by-id?key={id}");
             if (!response.IsSuccessStatusCode) return NotFound();
 
@@ -29,7 +39,19 @@ namespace eBookStore.Pages.Publishers
 
         public async Task<IActionResult> OnPostAsync()
         {
+            var token = Request.Cookies["Token"];
+
+            if (string.IsNullOrEmpty(token))
+            {
+                return Unauthorized(); // Or redirect to login page
+            }
+            // Add the Authorization header with Bearer token
+            _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
             var response = await _httpClient.DeleteAsync($"Publishers?key={Publisher.pub_id}");
+            if (response.StatusCode == HttpStatusCode.Forbidden)
+            {
+                return RedirectToPage("/AccessDenied");
+            }
 
             if (!response.IsSuccessStatusCode) return Page();
